fix: added clientID to server response
Added the clientID of the client who requested the usernames to the response. Therefore, the previous usernames array was converted to an ArrayList that contains the args to be send, where the first element is the clientID. When sending those args, they are converted to an array, to match the broadcast-method signature.