feat: added method argument - ceareteSelectionPopup
This method now requires an ArrayList that contains the usernames of the players in the game. They should be sorted by their id, as they will be put next to a figure, that should match their actual figure. If the given usernames are not valid, default names will be used.