refactor: changed variable-names to fulfill SQG
Renamed the prefix variable and the argument arrays, as SonarCloud flagged this as a duplicate. However, those variables are necessary to parse the given arguments.

Theoretically the array.length == 0 check could be removed. However, as this would be the results of a invalid call, thus making the call useless, a log-statement has been introduced.