fix: fixed condition for future prefix removal
Fixed a bug in which the arguments where set to the prefix when the message had a format like this: "PREFIX 1". This is because in this case the  args array had a length of 1 as there was no semicolon in the message. Now the check on whether the array only contains one argument like "1", has been updated.