Fixed a nullreference bug in the GameMap creation process
When initializing the GameMap object , it would instantiate SpecialFields of Type RiskTaskFields which would in turn get assigned a fixed Card (not correct, because for every field the card would remain the same throughout the game) which in turn references the gamemap instance to get the position of certain fields with a given Name. This was fixed by Changing the Risktask class and instead assigning each RiskTaskField a property that holds a reference to the Tast-Card-Stapel. Whenever a card is needed the TaskCollection Class has a public method to draw a random Card from the stapel